Add files via upload
The tweet archive of the Twitter user @dog rates, commonly known as WeRateDogs, is the subject of this dataset. This account rate people’s dogs and comments accordingly. To start the data wrangling and analysis process, the required datasets were gathered using the request library and via the Twitter API, and then assessed visually and programmatically. The datasets were later cleaned to remove unclean and untidy data. To produce the necessary results, an exploratory data analysis (EDA) was conducted using matplotlib and seaborn.